requirements.txt
alchemy==20.5
cachelib==0.2.0
certifi==2021.5.30
chardet==3.0.4
click==8.0.1
docutils==0.17.1
filelock==3.0.12
Flask==2.0.1
Flask-Session==0.4.0
greenlet==1.1.0
idna==2.8
importlib-metadata==4.6.1
itsdangerous==2.0.1
Jinja2==3.0.1
lockfile==0.12.2
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
python-daemon==2.2.4
requests==2.22.0
SQLAlchemy==1.4.22
typing-extensions==3.10.0.0
urllib3==1.25.11
Werkzeug==2.0.1
zipp==3.5.0
